fixed fload issues Enabled Cart Clear on sale

+++ backend/scripts/register.py
@@ -1,4 +1,5 @@
 import datetime
+from decimal import Decimal
 from os.path import exists
 from models.RegisterItem import Item
 import time
@@ -92,14 +93,14 @@ class Register:
         if not items:
             raise NoItemsError
         cart = ''
-        total: float = 0
-        tax: float = items[0].tax
+        total: Decimal = 0
+        tax: Decimal = items[0].tax
         # generate item string
         for item in items:
             # check for correct tax rate
             if item.tax != tax:
                 raise InconsistentTaxError
-            taxFactor: float = 1.0 + tax
+            taxFactor: Decimal = Decimal(1.0) + tax
             taxPrecent = f"{int(tax * 100)}%"
 
             # line length = 42 chars -> 3 margin, 9 cnt, 17 name, 10 price, 3 margin
@@ -110,7 +111,7 @@ class Register:
             cart += to_hex('   ') + '0D'
 
             # increase cart value
-            total += float(item.price * item.count)
+            total += item.price * item.count
 
         self._transmit(
             # start flashing to memory
